fix: remove calling of createConfirmDeviceCredentialIntent intent
Lots of information regarding this can be found in
comments of the referenced issue.

Long story short it seems that the unlock enforced by
the createConfirmDeviceCredentialIntent intent has no
effect in the keys created by a KeyPairGeneratorSpec.

Additionally the feature that required the UNLOCK
intent seems to have no effect in the latest versions
of Android, which includes Android 10.

Not calling anything at all seems to not have an issue.

References https://outsystemsrd.atlassian.net/browse/RNMT-3803